:art: refactor: ParkingMap 컴포넌트 분리
- ParkingMap 컴포넌트를 src/app/components/ParkingMap.tsx 파일로 이동합니다.
